requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
paramiko==3.3.1
twisted==23.8.0
pyftpdlib==1.5.7
elasticsearch==8.10.0
requests==2.31.0
geoip2==4.7.0
pyyaml==6.0.1
python-socketio==5.9.0
eventlet==0.33.3
psutil==5.9.6
cryptography==41.0.7
scapy==2.5.0
maxminddb>=2.3.0
python-dateutil==2.8.2
colorama==0.4.6
rich==13.6.0
click==8.1.7
schedule==1.2.0
jinja2==3.1.2
werkzeug==2.3.7
itsdangerous==2.1.2
markupsafe==2.1.3
six==1.16.0
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4